pyproject.toml
[project]
name = "alemno-backend"
version = "0.1.0"
description = "Alemno Backend API for customer and loan management"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.6",
    "djangorestframework>=3.15.0",
    "psycopg[binary]>=3.1.0",
]
